+++ modules/techniques/techniques.py
@@ -261,7 +261,7 @@ def generate_data_for_md(technique_dict, technique, tactic_list, is_sub_techniqu
             # Get data sources
             if technique.get('x_mitre_data_sources'):
                 technique['x_mitre_data_sources'].sort()
-                technique_dict['data_sources'] = ", ".join(technique['x_mitre_data_sources'])
+                technique_dict['data_sources'] = data_sources_link(technique['x_mitre_data_sources'])
 
             # Get if technique supports remote
             if technique.get('x_mitre_remote_support'):
@@ -555,3 +555,68 @@ def get_subtechniques(technique):
                 subtechs.append(sub_data)
     
     return sorted(subtechs, key=lambda k: k['id'])
+
+new_data_source_mapping = {
+    "Active Directory":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/active_directory.yml",
+    "Application Log":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/application_log.yml",
+    "Cloud Service":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/cloud_service.yml",
+    "Cloud Storage":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/cloud_storage.yml",
+    "Cluster":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/cluster.yml",
+    "Command":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/command.yml",
+    "Container":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/container.yml",
+    "Drive":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/drive.yml",
+    "Driver":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/driver.yml",
+    "File":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/file.yml",
+    "Firewall":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/firewall.yml",
+    "Firmware":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/firmware.yml",
+    "Group":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/group.yml",
+    "Image":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/image.yml",
+    "Instance":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/instance.yml",
+    "Kernel":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/kernel.yml",
+    "Logon Session":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/logon_session.yml",
+    "Module":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/module.yml",
+    "Named Pipe":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/named_pipe.yml",
+    "Network Share":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/network_share.yml",
+    "Network Traffic":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/network_traffic.yml",
+    "Pod":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/pod.yml",
+    "Process":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/process.yml",
+    "Scheduled Job":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/scheduled_job.yml",
+    "Script":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/script.yml",
+    "Sensor Health":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/sensor_health.yml",
+    "Service":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/service.yml",
+    "Snapshot":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/snapshot.yml",
+    "User Account":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/user_account.yml",
+    "Volume":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/volume.yml",
+    "Web Credential":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/web_credential.yml",
+    "Windows Registry":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/windows_registry.yml",
+    "WMI":"https://github.com/mitre-attack/attack-datasources/blob/main/contribution/wmi_object.yml"
+}
+
+def data_sources_link(data_sources):
+    """ Data source mapping for April release """
+
+    link = "<a target='_blank' href='{}'>{}</a>"
+
+    def replace_link(data_sources):
+        
+        updated_data_sources = []
+
+
+        for data_source in data_sources:
+            data_source_str = data_source
+            if ":" in data_source:
+                data_source_name = data_source.split(":")[0].strip()
+                data_source_component = data_source.split(":")[1].strip()
+
+                if new_data_source_mapping.get(data_source_name):
+                    data_source_str = f"{link.format(new_data_source_mapping.get(data_source_name), data_source_name)}: {data_source_component}"         
+            else:
+                if new_data_source_mapping.get(data_source):
+                    data_source_str = f"{link.format(new_data_source_mapping.get(data_source), data_source)}" 
+
+            updated_data_sources.append(data_source_str)
+        
+        return updated_data_sources
+    
+    data_sources.append("Command: Command Execution")
+    return ", ".join(replace_link(data_sources))