+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Added navigator support for mitigations and updated navigator layer support to 4.2. See issue [#234](https://github.com/mitre-attack/attack-website/issues/234).
 - Added ATT&CK IDs to Mitigations and Procedure Examples on technique pages, and software and group home pages. See issues [#235](https://github.com/mitre-attack/attack-website/issues/235)[#236](https://github.com/mitre-attack/attack-website/issues/236).
 - Improved the [Working with ATT&CK](https://attack.mitre.org/resources/working-with-attack/) page with additional information about ATT&CK data and the tools with which it can be manipulated. The new page also includes generated Excel spreadsheets representing the ATT&CK knowledge base (see also [mitreattack-python](https://github.com/mitre-attack/mitreattack-python)'s attackToExcel converter). See issue [#142](https://github.com/mitre-attack/attack-website/issues/142).
+- Data sources in technique cards are now hyperlinks to the [attack-datasources GitHub repository](https://github.com/mitre-attack/attack-datasources).
 
 # 23 March 2021
 ## ATT&CK website version 3.1.1